refactor: :recycle: function naming convention
