+++ frontend/src/common/utils/menuMapper/constants.ts
