+++ frontend/src/common/utils/menuMapper/types.ts
